4.22: Bloglist expansion, step10

+++ part4/blog/backend/controllers/blogs.js
@@ -48,8 +48,22 @@ blogsRouter.post('/', async (request, response) => {
   }
 })
 blogsRouter.delete('/:id', async (request, response) => {
-  await Blog.findByIdAndRemove(request.params.id)
-  response.status(204).end()
+  const decodedToken = jwt.verify(request.token, process.env.SECRET)
+  if (!decodedToken.id) {
+    return response.status(401).json({ error: 'token missing or invalid' })
+  }
+  const blog = await Blog.findById(request.params.id)
+  const user = await User.findById(decodedToken.id)
+  //console.log(`${blog} and user is ${user}`)
+  if (!blog) {
+    return response.status(404).json({ error: 'No such blog found' })
+  }
+  if (blog.user.toString() === user.id.toString()) {
+    await Blog.findByIdAndRemove(request.params.id)
+    response.status(204).end()
+  } else {
+    response.status(403).json({ error: 'no rights to perform the delete action' })
+  }
 })
 blogsRouter.put('/:id', async (request, response) => {
   const body = request.body
